+++ models/base_model.py
@@ -8,19 +8,6 @@ import functools
 import operator
 
 
-class BaseFactory(ABC):
-    def __init__(self, module_name_down, module_name_up, modules_lib):
-        self.module_name_down = module_name_down
-        self.module_name_up = module_name_up
-        self.modules_lib = modules_lib
-
-    def get_module(self, index, flow):
-        if flow.upper() == "UP":
-            return getattr(self.modules_lib, self.module_name_up, None)
-        else:
-            return getattr(self.modules_lib, self.module_name_down, None)
-
-
 class BaseModel(torch.nn.Module):
     """This class is an abstract base class (ABC) for models.
     To create a subclass, you need to implement the following five functions:
